+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `convert` function modified
+- Test system modified
 ## [1.1] - 2025-11-10
 ### Added
 - English circled mode